[MOD] Traducciones y pequeñas modificaciones
