+++ mrp_production_notes/i18n/mrp_production_notes.pot
@@ -1,0 +1,28 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* mrp_production_notes
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 8.0rc1\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2014-08-04 08:50+0000\n"
+"PO-Revision-Date: 2014-08-04 08:50+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: mrp_production_notes
+#: model:ir.model,name:mrp_production_notes.model_mrp_production
+msgid "Manufacturing Order"
+msgstr ""
+
+#. module: mrp_production_notes
+#: view:mrp.production:0
+#: field:mrp.production,notes:0
+msgid "Notes"
+msgstr ""
+
