+++ stock_lot_notes/i18n/es.po
@@ -1,0 +1,28 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* stock_lot_notes
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 8.0rc1\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2014-08-04 08:51+0000\n"
+"PO-Revision-Date: 2014-08-04 10:52+0100\n"
+"Last-Translator: Oihane <oihanecruce@gmail.com>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: \n"
+
+#. module: stock_lot_notes
+#: view:stock.production.lot:0
+#: field:stock.production.lot,notes:0
+msgid "Notes"
+msgstr "Notas"
+
+#. module: stock_lot_notes
+#: model:ir.model,name:stock_lot_notes.model_stock_production_lot
+msgid "Serial Number"
+msgstr "Nº de serie"
+
