+++ stock_lot_notes/views/lot_view.xml
@@ -7,12 +7,11 @@
             <field name="inherit_id" ref="stock.view_production_lot_form"/>
             <field name="arch" type="xml">
                 <notebook position="inside">
-                    <page string="Remark">
-                        <field name="remark" nolabel="1"/>
+                    <page string="Notes">
+                        <field name="notes" nolabel="1"/>
                     </page>
                 </notebook>
             </field>
         </record>
-    
     </data>
 </openerp>
